requirements/runtime.txt
Django >= 2.0
django-bootstrap4
django-tables2
djangorestframework
